+++ configs/test-utils/src/CustomNotError.ts
@@ -1,5 +1,0 @@
-export class CustomNotError {
-  constructor(public message?: string) {
-    console.log(message)
-  }
-}