+++ packages/react/src/index.ts
@@ -4,12 +4,9 @@ export { ErrorBoundary, useErrorBoundary, useErrorBoundaryFallbackProps } from '
 export { ErrorBoundaryGroup, useErrorBoundaryGroup } from './ErrorBoundaryGroup'
 export { Delay } from './Delay'
 export { wrap } from './wrap'
+export { DevMode } from './DevMode'
 
-// type
 export type { SuspenseProps } from './Suspense'
 export type { ErrorBoundaryProps, ErrorBoundaryFallbackProps } from './ErrorBoundary'
 export type { ErrorBoundaryGroupProps } from './ErrorBoundaryGroup'
 export type { DelayProps } from './Delay'
-
-// experimental
-export { DevMode } from './DevMode'